+++ packages/app/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.3](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.11.2...v2.11.3) (2026-09-02)
+
+### Bug Fixes
+
+- **app:** anchor the notes button to the thumb edge and center the tools row ([#387](https://github.com/vitalyiegorov/suuudokuuu/issues/387)) ([bb667da](https://github.com/vitalyiegorov/suuudokuuu/commit/bb667da28f4743399bd460714f6fc0904d6089d5))
+
+## [2.11.2](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.11.1...v2.11.2) (2026-09-01)
+
+**Note:** Version bump only for package @suuudokuuu/app
+
+## [2.11.1](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.11.0...v2.11.1) (2026-09-01)
+
+**Note:** Version bump only for package @suuudokuuu/app
+
+# [2.11.0](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.10.1...v2.11.0) (2026-08-31)
+
+### Features
+
+- **app:** hide undo and redo in hardcore mode with e2e coverage ([#364](https://github.com/vitalyiegorov/suuudokuuu/issues/364)) ([4b208b1](https://github.com/vitalyiegorov/suuudokuuu/commit/4b208b145524e70e5f9c3a5f7169ee450100ab0a))
+
 ## [2.10.1](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.10.0...v2.10.1) (2026-08-30)
 
 ### Bug Fixes